Move these constants to _devkit_constants
