+++ src/promptflow/promptflow/_constants.py
@@ -164,11 +164,3 @@ class MessageFormatType:
 
 
 DEFAULT_OUTPUT_NAME = "output"
-
-OUTPUT_FILE_NAME = "output.jsonl"
-
-
-class OutputsFolderName:
-    FLOW_OUTPUTS = "flow_outputs"
-    FLOW_ARTIFACTS = "flow_artifacts"
-    NODE_ARTIFACTS = "node_artifacts"
