+++ src/promptflow/tests/executor/e2etests/test_logs.py
@@ -3,7 +3,7 @@ from tempfile import mkdtemp
 
 import pytest
 
-from promptflow._constants import OUTPUT_FILE_NAME
+from promptflow._devkit_constants import OUTPUT_FILE_NAME
 from promptflow._utils.logger_utils import LogContext
 from promptflow.batch import BatchEngine
 from promptflow.batch._result import BatchResult
